refactor: move the Clock and conference-timezone providers to :core:common
Auditing what still blocks the next extraction turned up the same shape of
problem @IoDispatcher had: SessionsViewModel injects @ConferenceTimeZone from
presentation/di, and home, sessions and MainViewModel all inject Clock. A
feature module cannot reach either without depending on :presentation, which
is the one thing a feature must not do.

Both providers move to :core:common as TimeModule. PresentationModule keeps
only the notification manager, which genuinely belongs with the app.

Every remaining feature area now imports nothing from :presentation. The one
exception is notifications, which needs MainActivity to build a PendingIntent
and is a service rather than a feature — it belongs with the composition root.
