+++ presentation/src/main/java/com/android254/presentation/sessions/view/SessionsViewModel.kt
@@ -21,7 +21,6 @@ import com.android254.domain.models.Session
 import com.android254.domain.repos.SessionsRepo
 import com.android254.domain.work.SyncDataWorkManager
 import com.android254.presentation.common.resultstatus.ResultStatus
-import com.android254.presentation.di.ConferenceTimeZone
 import com.android254.presentation.models.EventDate
 import com.android254.presentation.models.SessionPresentationModel
 import com.android254.presentation.models.SessionsFilterCategory
@@ -30,6 +29,7 @@ import com.android254.presentation.sessions.mappers.toPresentationModel
 import com.android254.presentation.sessions.models.SessionsIntentHandler
 import com.android254.presentation.sessions.models.SessionsUiState
 import dagger.hilt.android.lifecycle.HiltViewModel
+import ke.droidcon.kotlin.core.common.di.ConferenceTimeZone
 import ke.droidcon.kotlin.core.common.di.IoDispatcher
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.PersistentList
